requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
email-validator==1.1.3
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
phonenumbers==8.12.45
platformdirs==2.5.1
PyJWT==2.3.0
python-decouple==3.6
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
urllib3==1.26.9
